raw_data_bronze/requirements.txt
yfinance
kafka-python
pyspark==3.5.7
